% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_protein.R
\name{annotate_protein}
\alias{annotate_protein}
\title{Combine Metadata with Protein-level Data}
\usage{
annotate_protein(data = data, sample_annotation = sample_annotation)
}
\arguments{
\item{data}{The data frame output from the 'prepare_data' function}

\item{sample_annotation}{The matrix output from the 'create_metadata' function}
}
\value{
A data frame containing protein-level MaxLFQ intensities and peptide count for each protein group
}
\description{
Annotates protein group-level data with metadata, imputes
missing values using a standard distribution as implemented in Perseus, and
counts the number of times a peptide was quantified
}
