% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{create_metadata}
\alias{create_metadata}
\title{Create Metadata Information}
\usage{
create_metadata(data)
}
\arguments{
\item{data}{The data frame output from the prepare_data function}
}
\value{
A matrix that has a condition name for every sample name
}
\description{
Creates metadata information, associating treatment groups to sample names
}
