% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_customer_plots.R
\name{make_customer_plots}
\alias{make_customer_plots}
\title{Make Customer Plots}
\usage{
make_customer_plots(
  pep_data = pep_data,
  pg_data = pg_data,
  t_test_data = t_test_data,
  filename = "Customer_report.pdf",
  sample_correlation = T,
  protein_coverage = T,
  protein_completeness = F,
  protein_intensity = F,
  volcano = T,
  fold_change_cutoff = 1,
  p_value_cutoff = 0.05
)
}
\arguments{
\item{pep_data}{The data frame generated from the 'annotate_peptide' function.}

\item{pg_data}{The data frame generated from the 'annotate_protein' function}

\item{t_test_data}{The data frame generated from the 'perform_t_test' function.}

\item{filename}{The desired name of the exported PDF. Should be in quotes and
end in .pdf}

\item{sample_correlation}{A heatmap displaying the spearman rank correlation
between each sample. Additionally perform hierarchical clustering between samples.}

\item{protein_coverage}{A bar graph depicting the number of proteins
quantified in a given sample.}

\item{protein_completeness}{A bar graph depicting the percentage of proteins
quantified in a given sample as compared to all proteins quantified.}

\item{protein_intensity}{A box and whisker plot over a violin plot depicting
the distribution of MaxLFQ protein abundances of each sample.}

\item{volcano}{A series of volcano plots For each group comparison. log2 fold
change are on the x-axis and log10 transformed p-values from the moderated t-
test are on the y-axis. By default, fold-change cut-off is set to 1 and p-value
significance level is set to 0.05. Any protein groups with values greater than
BOTH cutoffs are highlioghted. Plots are faceted by condition.}

\item{fold_change_cutoff}{The log2 fold change you want to have the volcano plot
cut-off at. Default is 1 (Corresponding to a 2x expression change between conditions).}

\item{p_value_cutoff}{The p-value you want to have the volcano plot cut-off at.
Default is 0.05.}
}
\value{
A PDF file written in your working directory
}
\description{
Exports a PDF containing a variety of plots meant primarily for
distribution to customers. All plots are adapted from their implementation in
the Protti R Package (https://cran.r-project.org/web/packages/protti/index.html).
By default, coloring for all plots is by condition groups and use the 'Acadia'
scheme in the NatParksPalettes R package
(https://github.com/kevinsblake/NatParksPalettes).
}
