% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qc_plots.R
\name{make_qc_plots}
\alias{make_qc_plots}
\title{Make QC Plots}
\usage{
make_qc_plots(
  pep_data = pep_data,
  pg_data = pg_data,
  filename = "QC_report.pdf",
  peak_shape = T,
  charge_state = T,
  cv = T,
  precursor_coverage = T,
  peptide_coverage = T,
  protein_coverage = T,
  precursor_completeness = T,
  peptide_completeness = T,
  protein_completeness = T,
  precursor_intensity = T,
  protein_intensity = T
)
}
\arguments{
\item{pep_data}{The data frame generated from the 'annotate_peptide' function.}

\item{pg_data}{The data frame generated from the 'annotate_protein' function}

\item{filename}{The desired name of the exported PDF. Should be in quotes and
end in .pdf}

\item{peak_shape}{A line plot depicting the mean peak width for each sample as
a function of retention time.}

\item{charge_state}{A bar graph depicting the percentage of precursor charge
states for every sample.}

\item{cv}{A box and whisker plot over a violin plot depicting the distribution
of protein-group MaxLFQ abundances among:
    all samples together
    all samples in each conditions}

\item{precursor_coverage}{A bar graph depicting the number of precursors
quantified in a given sample.}

\item{peptide_coverage}{A bar graph depicting the number of peptides (can
include multiple charge states) quantified in a given sample.}

\item{protein_coverage}{A bar graph depicting the number of proteins
quantified in a given sample.}

\item{precursor_completeness}{A bar graph depicting the percentage of precursors
quantified in a given sample as compared to all precursors quantified.}

\item{peptide_completeness}{A bar graph depicting the percentage of peptides
(can include multiple charge states) quantified in a given sample as compared
to all peptides quantified.}

\item{protein_completeness}{A bar graph depicting the percentage of proteins
quantified in a given sample as compared to all proteins quantified.}

\item{precursor_intensity}{A box and whisker plot over a violin plot depicting
the distribution of normalized precursor intensities of each sample.}

\item{protein_intensity}{A box and whisker plot over a violin plot depicting
the distribution of MaxLFQ protein abundances of each sample.}
}
\value{
A PDF file written in your working directory.
}
\description{
Exports a PDF containing a variety of plots meant primarily for
internal QC purposes to quickly assess quality of the data. All plots are adapted
from their implementation in the Protti R Package
(https://cran.r-project.org/web/packages/protti/index.html). By default,
coloring for all plots is by condition groups and use the 'Acadia' scheme in the
NatParksPalettes R package (https://github.com/kevinsblake/NatParksPalettes).
}
