% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_protein.R
\name{pep_count_fun}
\alias{pep_count_fun}
\title{Combine Metadata with Protein-level Data}
\usage{
pep_count_fun(df)
}
\value{
A data frame containing protein-level MaxLFQ intensities and peptide count for each protein group
}
\description{
Annotates protein group-level data with metadata and counts the number of times a peptide was quantified
}
