% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_t_test.R
\name{perform_t_test}
\alias{perform_t_test}
\title{Perform Moderated T-Test}
\usage{
perform_t_test(pg_data = pg_data)
}
\arguments{
\item{pg_data}{Data frame output from the 'annotate_protein' function}
}
\value{
A data frame containing the log2 fold change between conditions for each protein group and its associated p-value.
}
\description{
Calculates differential abundance and performs a moderated
t-test as implemented in the protti R package (https://cran.r-project.org/web/packages/protti/index.html).
Calculations are performed on protein-group abundances and between all possible
comparisons.
}
