% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Import and Filter DIA-NN data and Calculate MaxLFQ}
\usage{
prepare_data(
  filepath = "report.tsv",
  contaminants = "Keratin",
  q = 0.01,
  group.q = 0.01,
  pep.value = 0.5
)
}
\arguments{
\item{filepath}{The file path (or name) or the main report.tsv file generated
by DIA-NN. The default uses is the file name given by DIA-NN in your active
directory.}

\item{contaminants}{Strings used to filter out proteins that are common
contaminants. The default is only Keratin}

\item{q}{The q value to use to filter precursor identifications.
Default is 0.01}

\item{group.q}{The q value to use to filter protein groupings.
Default is 0.01}

\item{pep.value}{The value to use to filter out proteins with higher posterior
error probabilities. Default is 0.50}
}
\value{
A data frame containing protein information, peptide sequence,
precursor normalized intensity, and protein group MaxLFQ
}
\description{
Performs 4 separate steps to prepare a DIA-NN output for data
analysis. 1 Data is imported as implemented in the diann package. 2 Precursors
are filtered. 3 According to protein groups, abundances are calculated using
the MaxLFQ algorithm as implemented in the diann package. 4 Data is tidied.
}
