% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_report.R
\name{write_report}
\alias{write_report}
\title{Generate Customer Report}
\usage{
write_report(
  t_test_data = t_test_data,
  pg_data = pg_data,
  filename = "output_report.tsv"
)
}
\arguments{
\item{t_test_data}{The data frame output from the 'perform_t_test' function.}

\item{pg_data}{The data frame output from the 'annotate_protein' function.}

\item{filename}{The desired name of the exported report. Should be in quotes
and end in .tsv}
}
\value{
A .tsv file written in your working directory.
}
\description{
Exports a .tsv into your working directory that can be further
edited and sent as a report to customers. Each row corresponds to a different
protein group and has rows for:
    MaxLFQ protein abundance of each sample (in log2)
    Number of peptides quantified for each sample
    Average MaxLFQ for each group
    Log2 fold-change of each group comparison
    p-value of each group comparison
}
